feat: merge crates in sel4-kit
