+++ Cargo.toml
@@ -8,8 +8,6 @@ members = [
   "crates/common",
   "crates/common-macros",
   "crates/sel4-runtime",
-  "crates/slot-manager",
-  "crates/task-helper",
   "crates/srv-gate",
   "services/blk-thread",
   # "services/fat-thread",
@@ -51,10 +49,8 @@ lwext4-thread = { path = "services/lwext4-thread" }
 
 # 一些公共包
 config = { path = "config" }
-task-helper = { path = "crates/task-helper" }
 common = { path = "crates/common" }
 common-macros = { path = "crates/common-macros" }
-slot-manager = { path = "crates/slot-manager" }
 srv-gate = { path = "crates/srv-gate" }
 flatten_objects = "0.2.2"
 zerocopy = { version = "0.8.20", features = ["alloc", "derive"] }
