+++ crates/common/Cargo.toml
@@ -9,7 +9,7 @@ edition = "2024"
 sel4 = { workspace = true }
 config = { workspace = true }
 common-macros = { workspace = true }
-slot-manager = { workspace = true }
+sel4-kit = { workspace = true }
 object = { version = "0.36.1", default-features = false, features = ["read"] }
 bitflags = "2.6"
 num_enum = { version = "0.7.3", default-features = false }
