--- crates/common/src/ipc/mod.rs
@@ -1,5 +1,0 @@
-//! 提供 IPC 相关的帮助函数
-//!
-//! 协助快速处理 IPC 消息
-#![deny(missing_docs)]
-pub mod ipc_saver;
