--- crates/drivers/pl011-driver/Cargo.toml
@@ -1,6 +1,0 @@
-[package]
-name = "pl011-driver"
-version = "0.1.0"
-edition.workspace = true
-
-[dependencies]
