--- crates/drivers/pl011-driver/src/lib.rs
@@ -1,1 +1,0 @@
-#![no_std]
