+++ crates/sel4-runtime/Cargo.toml
@@ -3,9 +3,6 @@ name = "sel4-runtime"
 version = "0.1.0"
 edition = "2024"
 
-[features]
-root-task = []
-
 [dependencies]
 sel4 = { workspace = true }
 sel4-ctors-dtors = { workspace = true }
@@ -19,6 +16,5 @@ sel4-runtime-common = { workspace = true, features = ["tls", "start"] }
 config = { workspace = true }
 common = { workspace = true }
 common-macros = { workspace = true }
-slot-manager = { workspace = true }
 sel4-kit = { workspace = true }
 log = "0.4"
