+++ crates/sel4-runtime/src/lib.rs
@@ -5,9 +5,7 @@
 #![deny(warnings)]
 #![deny(missing_docs)]
 
-#[cfg(not(feature = "root-task"))]
 pub mod entry;
-#[cfg(not(feature = "root-task"))]
 pub mod heap;
 pub mod logging;
 pub mod macros;
