--- crates/slot-manager/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "slot-manager"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-# rust-sel4 相关依赖
-sel4 = { workspace = true }
