--- crates/task-helper/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "task-helper"
-version = "0.1.0"
-edition = "2024"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-config = { workspace = true }
-sel4 = { workspace = true }
-sel4-sync = { workspace = true }
-common = { workspace = true }
-xmas-elf = "0.9.1"
-log = "0.4"
-spin = { workspace = true }
-slot-manager = { workspace = true }
