+++ root-task/Cargo.toml
@@ -11,21 +11,19 @@ sel4-elf-header = { workspace = true }
 sel4-logging = { workspace = true }
 sel4-initialize-tls = { workspace = true }
 sel4-stack = { workspace = true }
-sel4-runtime = { workspace = true, features = ["root-task"] }
+sel4-sync = { workspace = true }
 
 sel4-kit = { workspace = true }
 # 自定义依赖
 config = { workspace = true }
 common = { workspace = true }
-slot-manager = { workspace = true }
-task-helper = { workspace = true }
 
 # 其他依赖
 spin = { workspace = true }
 log = "0.4"
 xmas-elf = "0.9.1"
 include_bytes_aligned = "0.1.3"
-object = { version = "0.36.1", default-features = false, features = ["read"] }
+object = { version = "0.36.5", default-features = false, features = ["read"] }
 
 [build-dependencies]
 serde = "1.0"
