+++ services/blk-thread/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2024"
 [dependencies]
 sel4 = { workspace = true, default-features = false }
 sel4-runtime = { workspace = true }
-slot-manager = { workspace = true }
+sel4-kit = { workspace = true }
 spin = { workspace = true }
 log = "0.4"
 common = { workspace = true }
