+++ services/lwext4-thread/Cargo.toml
@@ -9,7 +9,7 @@ sel4-runtime = { workspace = true }
 spin = { workspace = true }
 log = "0.4"
 common = { workspace = true }
-slot-manager = { workspace = true }
+sel4-kit = { workspace = true }
 
 hashbrown = { workspace = true }
