+++ services/uart-thread/Cargo.toml
@@ -11,7 +11,7 @@ spin = { workspace = true }
 log = "0.4"
 config = { workspace = true }
 common = { workspace = true }
-slot-manager = { workspace = true }
+sel4-kit = { workspace = true }
 arm_pl011 = { git = "https://github.com/Byte-OS/arm_pl011.git", rev = "8a66e24" }
 futures = { version = "0.3.28", default-features = false, features = ["alloc"] }
 srv-gate = { workspace = true }
